pyproject.toml
[project]
name = "final-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.35.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.6.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
